Add backends tests
